pyproject.toml
[tool.poetry]
name = "rag-chatbot-langchain-railway"
version = "0.1.0"
description = ""
authors = ["Cristian Encalada"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.12"
langchain-cli = "^0.0.21"
langserve = "^0.0.51"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.28"
tavily-python = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
